requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.8.30=hbcca054_0
contextlib2=21.6.0=pypi_0
filelock=3.13.1=pypi_0
fsspec=2024.2.0=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.2=pypi_0
ld_impl_linux-64=2.43=h712a8e2_1
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
libnsl=2.0.1=hd590300_0
libsqlite=3.47.0=hadc24fc_0
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
markupsafe=2.1.5=pypi_0
ml-collections=0.1.1=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=he02047a_1
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=9.1.0.70=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.21.5=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.3.2=hb9d3cd8_0
pillow=10.2.0=pypi_0
pip=24.2=pyh8b19718_1
python=3.9.20=h13acc7a_1_cpython
pyyaml=6.0.2=pypi_0
readline=8.2=h8228510_1
scikit-learn=1.5.2=pypi_0
scipy=1.13.1=pypi_0
setuptools=75.1.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
sympy=1.13.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=2.5.0+cu118=pypi_0
torchaudio=2.5.0+cu118=pypi_0
torchvision=0.20.0+cu118=pypi_0
triton=3.1.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024b=hc8b5060_0
wheel=0.44.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
